requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
langchain-community>=0.2.0
langchainhub>=0.1.15
openai>=1.30.0
chromadb>=0.5.0
pypdf>=4.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
